backend/notifications/requirements.txt
requests==2.27.1
APScheduler==3.9.1
python-dotenv==0.20.0